pyproject.toml
[project]
name = "cafe-management-m"
version = "0.1.0"
description = ""
authors = [
    {name = "AleksandrZaec",email = "zaecam95@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-crispy-forms (>=2.3,<3.0)",
    "crispy-bootstrap5 (>=2024.10,<2025.0)",
    "django-filter (>=24.3,<25.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
